Added IIIT-B
Hi Rico! :)

We are using Docpress at IIIT-Bangalore to make course material available to students. This website was for us to test the waters and we are loving it so far and will most definitely be using it with more courses.

Thank you for making Docpress available :)

Regards,
Aravind Reddy V.